ActiveRaceSocket fixes
