+++ info-screens/src/racetrack/services/sockets/ActiveRaceSocket.ts
@@ -1,19 +1,49 @@
-import { io, Socket } from "socket.io-client";
+import { io, Socket } from 'socket.io-client';
+import { useState } from 'react';
+import { ActiveRace, RaceState } from 'state';
+import { RaceStateChanged } from 'services/racecontrol.service';
 
-const SOCKET_URL = "http://localhost:3001"; // Change if needed
+// This file will manage the socket connection and handle state updates
+export const useRaceSocket = () => {
+  const [activeRace, setActiveRace] = useState<ActiveRace>({
+    contestants: [],
+    raceState: 'pending',
+    isHazard: false,
+    isDanger: false,
+    isFinishing: false,
+    timeLeft: '00:00',
+  });
 
-let socket: Socket;
+  let socket: Socket;
 
-export const connectSocket = () => {
-    if (!socket) {
-        socket = io(SOCKET_URL);
-    }
-};
+  const connectSocket = () => {
+    // Create a new socket connection
+    socket = io("http://localhost:3001");
 
-export const disconnectSocket = () => {
+    socket.on("raceStateChanged", (raceState: RaceStateChanged) => {
+      console.log("Race state changed:", raceState.newState);
+
+      const fixedRaceState = Array.isArray(raceState.newState) 
+        ? raceState.newState[0] 
+        : raceState.newState;
+
+      setActiveRace({
+        ...activeRace,
+        raceState: fixedRaceState as RaceState,
+        isHazard: raceState.isHazard ?? false,
+        isDanger: raceState.isDanger ?? false,
+        isFinishing: raceState.isFinishing ?? false,
+        timeLeft: "00:00",
+      });
+    });
+  };
+
+  const disconnectSocket = () => {
+    // Logic to disconnect socket when component unmounts or no longer needed
     if (socket) {
         socket.disconnect();
     }
-};
+  };
 
-export { socket };
+  return { activeRace, connectSocket, disconnectSocket };
+};
